+++ src/main/java/site/bitrun/cryptocurrency/constants/AppConst.java
@@ -1,0 +1,6 @@
+package site.bitrun.cryptocurrency.constants;
+
+public abstract class AppConst {
+
+    public static final long MEMBER_STARTING_ASSET = 10_000_000; // 회원 가입 초기 자금
+}
