+++ src/main/java/site/bitrun/cryptocurrency/constants/session/SessionConst.java
@@ -1,4 +1,4 @@
-package site.bitrun.cryptocurrency.session;
+package site.bitrun.cryptocurrency.constants.session;
 
 public abstract class SessionConst {
     public static final String LOGIN_MEMBER = "loginMember";
